NativePixmapDmaBuf: Fix type disagreements.
Bug: 1411749
Test: None
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4209792
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1103405}

+++ ui/gfx/linux/native_pixmap_dmabuf.cc
@@ -35,17 +35,17 @@ int NativePixmapDmaBuf::GetDmaBufFd(size_t plane) const {
 
 uint32_t NativePixmapDmaBuf::GetDmaBufPitch(size_t plane) const {
   DCHECK_LT(plane, handle_.planes.size());
-  return handle_.planes[plane].stride;
+  return base::checked_cast<uint32_t>(handle_.planes[plane].stride);
 }
 
 size_t NativePixmapDmaBuf::GetDmaBufOffset(size_t plane) const {
   DCHECK_LT(plane, handle_.planes.size());
-  return static_cast<size_t>(handle_.planes[plane].offset);
+  return base::checked_cast<size_t>(handle_.planes[plane].offset);
 }
 
 size_t NativePixmapDmaBuf::GetDmaBufPlaneSize(size_t plane) const {
   DCHECK_LT(plane, handle_.planes.size());
-  return static_cast<size_t>(handle_.planes[plane].size);
+  return base::checked_cast<size_t>(handle_.planes[plane].size);
 }
 
 uint64_t NativePixmapDmaBuf::GetBufferFormatModifier() const {
